apps/api/pyproject.toml
[project]
name = "betterlabs-api"
version = "0.0.1"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "python-dotenv>=1.0",
  "pydantic>=2.8",
  "openai>=1.50",
  "httpx>=0.27",
  "fpdf2>=2.7.9",
  "markdown>=3.6",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
